\name{preprocessMOT-package}
\alias{preprocessMOT-package}
\alias{preprocessMOT}
\docType{package}
\title{
A package for pre-processing of MOTassinger data files.
}
\description{
This package contains scripts used for pre-processing of MOTassinger data files. Ideally suited for preprocessing of raw Affymetrix .CEL files.
However, this package can perform initial quality control (data distribution, sample-sample correlation, PCA) of data derived from any platform. Please note that the R codes used in this package have been adapted from various sources.
}
\details{
\tabular{ll}{
Package: \tab preprocessMOT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-03-03\cr
License: \tab GPL-2 \cr
}
	QC, check batch effects, sample sample correlation
}
\author{
Pawan Poudel

Maintainer: Pawan Poudel, paawanpoudel@gmail.com
}
\references{
	This package is the part of the MOTassigner Sadanandam et al. 2016 paper, which performs QC of the dataset used in the paper.
}

\keyword{package}
